+++ src/clj_kondo/core.clj
@@ -75,6 +75,8 @@
 
   - `:parallel`: optional. A boolean indicating if sources should be linted in parallel.`
 
+  - `:copy-configs`: optional. A boolean indicating if scanned hooks should be copied to clj-kondo config dir.`
+
   Returns a map with `:findings`, a seqable of finding maps, a
   `:summary` of the findings and the `:config` that was used to
   produce those findings. This map can be passed to `print!` to print
@@ -88,7 +90,9 @@
            :config
            :config-dir
            :parallel
-           :no-warnings]
+           :no-warnings
+           :dependencies
+           :copy-configs]
     :or {cache true}}]
   (let [start-time (System/currentTimeMillis)
         cfg-dir
@@ -117,7 +121,8 @@
         used-nss (atom {:clj #{}
                         :cljs #{}
                         :cljc #{}})
-        ctx {:no-warnings no-warnings
+        ctx {:dependencies (or dependencies no-warnings)
+             :copy-configs copy-configs
              :config-dir cfg-dir
              :config config
              :classpath classpath
@@ -145,10 +150,10 @@
         idacs (core-impl/index-defs-and-calls ctx)
         idacs (cache/sync-cache idacs cache-dir)
         idacs (overrides idacs)
-        _ (when (and no-warnings (not analysis))
+        _ (when (and dependencies (not analysis))
             ;; analysis is called from lint-var-usage, this can probably happen somewhere else
             (l/lint-var-usage ctx idacs))
-        _ (when-not no-warnings
+        _ (when-not dependencies
             (l/lint-var-usage ctx idacs)
             (l/lint-unused-namespaces! ctx)
             (l/lint-unused-private-vars! ctx)