+++ src/clj_kondo/impl/analyzer.clj
@@ -424,7 +424,7 @@
           :list exprs
           (recur rest-exprs))))))
 
-(defn analyze-defn [ctx expr]
+(defn analyze-defn [ctx expr defined-by]
   (let [ns-name (-> ctx :ns :name)
         ;; "my-fn docstring" {:no-doc true} [x y z] x
         [name-node & children] (next (:children expr))
@@ -494,6 +494,7 @@
                    :macro macro?
                    :private private?
                    :deprecated deprecated
+                   :defined-by defined-by
                    :fixed-arities (not-empty fixed-arities)
                    :arglist-strs (when (:analyze-arglists? ctx)
                                    arglist-strs)
@@ -848,7 +849,7 @@
 
 (declare analyze-defmethod)
 
-(defn analyze-def [ctx expr]
+(defn analyze-def [ctx expr defined-by]
   ;; (def foo ?docstring ?init)
   (let [children (next (:children expr))
         var-name-node (->> children first (meta/lift-meta-content2 ctx))
@@ -861,14 +862,15 @@
                           var-name
                           expr
                           (assoc-some metadata
-                                      :doc docstring)))
+                                      :doc docstring
+                                      :defined-by defined-by)))
     (analyze-children (assoc ctx
                              :in-def var-name)
                       (nnext (:children expr)))))
 
 (declare analyze-defrecord)
 
-(defn analyze-schema [ctx fn-sym expr]
+(defn analyze-schema [ctx fn-sym expr defined-by]
   (let [{:keys [:expr :schemas]}
         (schema/expand-schema ctx
                               fn-sym
@@ -876,17 +878,17 @@
     (concat
      (case fn-sym
        fn (analyze-fn ctx expr)
-       def (analyze-def ctx expr)
-       defn (analyze-defn ctx expr)
+       def (analyze-def ctx expr defined-by)
+       defn (analyze-defn ctx expr defined-by)
        defmethod (analyze-defmethod ctx expr)
-       defrecord (analyze-defrecord ctx expr 'defrecord))
+       defrecord (analyze-defrecord ctx expr defined-by))
      (analyze-children ctx schemas))))
 
 (defn analyze-binding-call [ctx fn-name binding expr]
   (let [callstack (:callstack ctx)
         config (:config ctx)
-        ns-name (-> ctx :ns :name)]
-    ;;(prn (:tag binding))
+        ns-name (-> ctx :ns :name)
+        fn-meta (meta fn-name)]
     (when-let [k (types/keyword binding)]
       (when-not (types/match? k :ifn)
         (findings/reg-finding! ctx (node->line (:filename ctx) expr :error
@@ -896,7 +898,11 @@
     (namespace/reg-used-binding! ctx
                                  ns-name
                                  binding
-                                 (meta expr))
+                                 (assoc (meta expr)
+                                        :name-row (:row fn-meta)
+                                        :name-col (:col fn-meta)
+                                        :name-end-row (:end-row fn-meta)
+                                        :name-end-col (:end-col fn-meta)))
     (when-not (config/skip? config :invalid-arity callstack)
       (let [filename (:filename ctx)
             children (:children expr)]
@@ -918,7 +924,7 @@
      ctx
      (node->line (:filename ctx) expr :warning :inline-def "inline def"))))
 
-(defn analyze-declare [ctx expr]
+(defn analyze-declare [ctx expr defined-by]
   (let [ns-name (-> ctx :ns :name)
         var-name-nodes (next (:children expr))]
     (doseq [var-name-node var-name-nodes]
@@ -926,7 +932,8 @@
                           (->> var-name-node (meta/lift-meta-content2 ctx) :value)
                           expr
                           (assoc (meta expr)
-                                 :declared true)))))
+                                 :declared true
+                                 :defined-by defined-by)))))
 
 (defn analyze-catch [ctx expr]
   (let [ctx (update ctx :callstack conj [nil 'catch])
@@ -1024,16 +1031,16 @@
 
 (defn analyze-defrecord
   "Analyzes defrecord, deftype and definterface."
-  [{:keys [:ns] :as ctx} expr resolved-as]
+  [{:keys [:ns] :as ctx} expr defined-by]
   (let [ns-name (:name ns)
         children (:children expr)
         children (next children)
         name-node (first children)
         name-node (meta/lift-meta-content2 ctx name-node)
         metadata (meta name-node)
-        metadata (assoc metadata :defined-by (symbol "clojure.core" (str resolved-as)))
+        metadata (assoc metadata :defined-by defined-by)
         record-name (:value name-node)
-        bindings? (not= 'definterface resolved-as)
+        bindings? (not= "definterface" (name defined-by))
         binding-vector (when bindings? (second children))
         field-count (when bindings? (count (:children binding-vector)))
         bindings (when bindings? (extract-bindings (assoc ctx
@@ -1044,13 +1051,13 @@
         arglists? (and bindings? (:analyze-arglists? ctx))
         ctx (ctx-with-bindings ctx bindings)]
     (namespace/reg-var! ctx ns-name record-name expr metadata)
-    (when-not (= 'definterface resolved-as)
+    (when-not (= "definterface" (name defined-by))
       (namespace/reg-var! ctx ns-name (symbol (str "->" record-name)) expr
                           (assoc-some metadata
                                       :arglist-strs (when arglists?
                                                       [(str binding-vector)])
                                       :fixed-arities #{field-count})))
-    (when (= 'defrecord resolved-as)
+    (when (= "defrecord" (name defined-by))
       (namespace/reg-var! ctx ns-name (symbol (str "map->" record-name))
                           expr (assoc-some metadata
                                            :arglist-strs (when arglists?
@@ -1314,7 +1321,7 @@
     (analyze-children ctx children false)))
 
 (defn analyze-call
-  [{:keys [:top-level? :base-lang :lang :ns :config :no-warnings] :as ctx}
+  [{:keys [:top-level? :base-lang :lang :ns :config :dependencies] :as ctx}
    {:keys [:arg-count
            :full-fn-name
            :row :col
@@ -1403,8 +1410,8 @@
                                                        :base-lang base-lang
                                                        :lang lang
                                                        :filename (:filename ctx)
-                                                       ;; save some memory during no-warnings
-                                                       :expr (when-not no-warnings expr)
+                                                       ;; save some memory during dependencies
+                                                       :expr (when-not dependencies expr)
                                                        :simple? (simple-symbol? full-fn-name)
                                                        :callstack (:callstack ctx)
                                                        :config (:config ctx)
@@ -1417,7 +1424,8 @@
                                                ns-name
                                                resolved-namespace)
                 (let [node expanded]
-                  (analyze-expression** ctx node)))
+                  (analyze-expression** (assoc-some ctx :defined-by (:defined-by transformed))
+                                        node)))
               ;;;; End macroexpansion
               (let [fq-sym (when (and resolved-namespace
                                       resolved-name)
@@ -1446,6 +1454,9 @@
                           (assoc ctx
                                  :resolved-as-clojure-var-name resolved-as-clojure-var-name)
                           ctx)
+                    defined-by (or (:defined-by ctx)
+                                   (when (and resolved-as-name resolved-as-namespace)
+                                     (symbol (name resolved-as-namespace) (name resolved-as-name))))
                     analyzed
                     (case resolved-as-clojure-var-name
                       ns
@@ -1455,16 +1466,16 @@
                                 (analyze-children ctx children))
                       alias
                       [(analyze-alias ctx expr)]
-                      declare (analyze-declare ctx expr)
+                      declare (analyze-declare ctx expr defined-by)
                       (def defonce defmulti goog-define)
                       (do (lint-inline-def! ctx expr)
-                          (analyze-def ctx expr))
+                          (analyze-def ctx expr defined-by))
                       (defn defn- defmacro definline)
                       (do (lint-inline-def! ctx expr)
-                          (analyze-defn ctx expr))
+                          (analyze-defn ctx expr defined-by))
                       defmethod (analyze-defmethod ctx expr)
                       defprotocol (analyze-defprotocol ctx expr)
-                      (defrecord deftype definterface) (analyze-defrecord ctx expr resolved-as-clojure-var-name)
+                      (defrecord deftype definterface) (analyze-defrecord ctx expr defined-by)
                       comment
                       (analyze-children ctx children)
                       (-> some->)
@@ -1528,19 +1539,18 @@
                         [clj-kondo.lint-as def-catch-all]
                         (analyze-def-catch-all ctx expr)
                         [schema.core fn]
-                        (analyze-schema ctx 'fn expr)
+                        (analyze-schema ctx 'fn expr 'schema.core/fn)
                         [schema.core def]
-                        (analyze-schema ctx 'def expr)
+                        (analyze-schema ctx 'def expr 'schema.core/def)
                         [schema.core defn]
-                        (analyze-schema ctx 'defn expr)
+                        (analyze-schema ctx 'defn expr 'schema.core/defn)
                         [schema.core defmethod]
-                        (analyze-schema ctx 'defmethod expr)
+                        (analyze-schema ctx 'defmethod expr 'schema.core/defmethod)
                         [schema.core defrecord]
-                        (analyze-schema ctx 'defrecord expr)
+                        (analyze-schema ctx 'defrecord expr 'schema.core/defrecord)
                         ([clojure.test deftest]
                          [cljs.test deftest])
-                        (test/analyze-deftest ctx expr
-                                              resolved-namespace resolved-name
+                        (test/analyze-deftest ctx expr defined-by
                                               resolved-as-namespace resolved-as-name)
                         [clojure.core.match match]
                         (match/analyze-match ctx expr)
@@ -1558,8 +1568,9 @@
                         (spec/analyze-fdef (assoc ctx
                                                   :analyze-children
                                                   analyze-children) expr)
-                        ([clojure.spec.gen.alpha lazy-combinators])
-                        (spec/analyze-lazy-combinators ctx expr)
+                        ([clojure.spec.gen.alpha lazy-combinators]
+                         [clojure.spec.gen.alpha lazy-prims])
+                        (analyze-declare ctx expr #_#_spec/analyze-lazy-combinators ctx expr)
                         [potemkin import-vars]
                         (potemkin/analyze-import-vars ctx expr)
                         ([clojure.core.async alt!] [clojure.core.async alt!!]
@@ -1568,6 +1579,8 @@
                                                         :analyze-expression** analyze-expression**
                                                         :extract-bindings extract-bindings)
                                                  expr)
+                        ([clojure.core.reducers defcurried])
+                        (analyze-defn ctx expr defined-by)
                         ([datahike.api q]
                          [datascript.core q]
                          [datomic.api q]
@@ -1634,7 +1647,7 @@
                                     :base-lang base-lang
                                     :lang lang
                                     :filename (:filename ctx)
-                                    :expr (when-not no-warnings expr)
+                                    :expr (when-not dependencies expr)
                                     :callstack (:callstack ctx)
                                     :config (:config ctx)
                                     :top-ns (:top-ns ctx)