+++ src/clj_kondo/impl/linters.clj
@@ -528,7 +528,14 @@
             :let [var-name (:name v)]
             :when (:private v)
             :when (not (contains? used-vars var-name))
-            :when (not (config/unused-private-var-excluded config ns-nm var-name))]
+            :when (not (config/unused-private-var-excluded config ns-nm var-name))
+            :when (not (utils/one-of (:defined-by v)
+                                     [clojure.core/defrecord
+                                      cljs.core/defrecord
+                                      clojure.core/deftype
+                                      cljs.core/deftype
+                                      clojure.core/definterface
+                                      cljs.core/definterface]))]
       (findings/reg-finding!
        ctx
        {:type :unused-private-var